refactorizacion de el directorio de componentes compartidos
