+++ src/app/(Recipes pages)/functions/GetAllRecipesOneCountry.ts
@@ -8,17 +8,17 @@ import { baseUrl } from '@libs/utils'
  */
 
 export async function GetAllRecipesOneCountry(id: string) {
-  try {
-    const response = await fetch(
-      `${baseUrl}/recipes/filter/countrys?countryId=${id}`
-    )
-    if (!response.ok) {
-      throw new Error('Error al obtener todas las recetas de 1 pais')
-    }
-    const data: Recipes[] = await response.json()
+	try {
+		const response = await fetch(
+			`${baseUrl}/recipes/filter/countrys?countryId=${id}`,
+		)
+		if (!response.ok) {
+			throw new Error('Error al obtener todas las recetas de 1 pais')
+		}
+		const data: Recipes[] = await response.json()
 
-    return data
-  } catch (error) {
-    console.error(error)
-  }
+		return data
+	} catch (error) {
+		console.error(error)
+	}
 }
