+++ src/app/(Recipes pages)/functions/GetRecipeById.ts
@@ -8,14 +8,14 @@ import { baseUrl } from '@libs/utils'
  */
 
 export async function GetRecipeById(id: string) {
-  try {
-    const response = await fetch(`${baseUrl}/recipes/${id}`)
-    if (!response.ok) {
-      throw new Error('Error al obtener  la receta')
-    }
-    const data: Recipes = await response.json()
-    return data
-  } catch (error) {
-    console.error(error)
-  }
+	try {
+		const response = await fetch(`${baseUrl}/recipes/${id}`)
+		if (!response.ok) {
+			throw new Error('Error al obtener  la receta')
+		}
+		const data: Recipes = await response.json()
+		return data
+	} catch (error) {
+		console.error(error)
+	}
 }
