+++ src/app/(Recipes pages)/paises/[id]/page.tsx
@@ -1,9 +1,9 @@
 import { GetAllRecipesOneCountry } from '@/(Recipes pages)/functions/GetAllRecipesOneCountry'
-import CardRecipe from '@/components/shared/CardRecipe'
-import DynamicTitle from '@/components/shared/DynamicTitle'
-import NotFound from '@/components/shared/NotFound'
-import ShowOneRecipe from '@/components/shared/ShowOneRecipe'
 import { type ParamProps } from '@interfaces/ParamProps.interface'
+import CardRecipe from '@shared-components/shared/CardRecipe'
+import DynamicTitle from '@shared-components/shared/DynamicTitle'
+import NotFound from '@shared-components/shared/NotFound'
+import ShowOneRecipe from '@shared-components/shared/ShowOneRecipe'
 import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import { Suspense } from 'react'
