+++ src/app/(Recipes pages)/recetas/components/oneRecipe/ShowDataRecipe.tsx
@@ -2,8 +2,8 @@ import { FormatRelativeDate } from '@/(Recipes pages)/functions/FormatRelativeDa
 import { GetRecipeById } from '@/(Recipes pages)/functions/GetRecipeById'
 import CheckBoxComponent from '@/(Recipes pages)/recetas/components/oneRecipe/CheckBox'
 import ChipComponent from '@/(Recipes pages)/recetas/components/oneRecipe/Chip'
-import { UiImage } from '@/components/ui'
 import GenerateJsonLD from '@libs/GenerateJsonLD'
+import { UiImage } from '@shared-components/ui'
 import { notFound } from 'next/navigation'
 import { FiList } from 'react-icons/fi'
 import BreadcrumbsComponent from './Breadcrumbs'
