+++ src/app/HomePage/components/sections/SectionCountry.tsx
@@ -1,5 +1,5 @@
 import RedirectPageCountry from '@/HomePage/components/RedirectPageCountry'
-import CardCountry from '@/components/shared/CardCountry'
+import CardCountry from '@shared-components/shared/CardCountry'
 import { GetAllCountry } from '../../functions/GetAllCountry'
 import ShowRecipesOneCountry from '../ShowRecipesOneCountry'
