+++ src/app/HomePage/components/skeletons/CountrySkeleton.tsx
@@ -1,4 +1,4 @@
-import SkeletonCardCountry from '@/components/skeletons/SkeletonCardCountry'
+import SkeletonCardCountry from '@shared-components/skeletons/SkeletonCardCountry'
 
 const CountrySkeleton = () => {
 	return (
