+++ src/app/HomePage/functions/GetAllCategories.ts
@@ -6,14 +6,14 @@ import { baseUrl } from '@libs/utils'
  * @returns Una promesa que resuelve en una lista de categorías (Categories) o muestra un error en la consola en caso de fallo.
  */
 export async function GetAllCategories() {
-  try {
-    const response = await fetch(`${baseUrl}/categories`)
-    if (!response.ok) {
-      throw Error('Error al obtener las categorias')
-    }
-    const data: Categories[] = await response.json()
-    return data
-  } catch (error) {
-    console.error(error)
-  }
+	try {
+		const response = await fetch(`${baseUrl}/categories`)
+		if (!response.ok) {
+			throw Error('Error al obtener las categorias')
+		}
+		const data: Categories[] = await response.json()
+		return data
+	} catch (error) {
+		console.error(error)
+	}
 }
