+++ src/app/HomePage/functions/GetLatestRecipes.ts
@@ -7,14 +7,14 @@ import { baseUrl } from '@libs/utils'
  */
 
 export async function GetLatestRecipes(limit: number) {
-  try {
-    const response = await fetch(`${baseUrl}/recipes/latest?limit=${limit}`)
-    if (!response.ok) {
-      throw new Error('Error al obtener las recetas')
-    }
-    const data: Recipes[] = await response.json()
-    return data
-  } catch (error) {
-    console.error(error)
-  }
+	try {
+		const response = await fetch(`${baseUrl}/recipes/latest?limit=${limit}`)
+		if (!response.ok) {
+			throw new Error('Error al obtener las recetas')
+		}
+		const data: Recipes[] = await response.json()
+		return data
+	} catch (error) {
+		console.error(error)
+	}
 }
