+++ src/components/navbar/Menu.tsx
@@ -10,18 +10,20 @@ interface Props {
 const Menu = ({ isOpenMenu, closeMenu, pathName }: Props) => {
 	return (
 		<ul
-			className={`Content-menu-items ${isOpenMenu ? 'opacity-100' : 'top-[-450px] opacity-0 md:opacity-100'
-				}`}
+			className={`Content-menu-items ${
+				isOpenMenu ? 'opacity-100' : 'top-[-450px] opacity-0 md:opacity-100'
+			}`}
 		>
 			{optionsMenu.map(({ href, title }, i) => (
 				<li key={title + i} className="m-3 md:m-auto p-2">
 					<Link
 						onClick={closeMenu}
 						href={href}
-						className={`Menu-item ${pathName === href
-							? 'text-gold md:border-b-2 lg:border-b-gold Active-link'
-							: 'text-slate-950'
-							}`}
+						className={`Menu-item ${
+							pathName === href
+								? 'text-gold md:border-b-2 lg:border-b-gold Active-link'
+								: 'text-slate-950'
+						}`}
 					>
 						{title}
 					</Link>
