+++ src/components/shared/CardRecipe.tsx
@@ -1,7 +1,7 @@
 import { formatDuration } from '@/(Recipes pages)/functions/FormatDurations'
 import ChipRecipesLatest from '@/HomePage/components/ChipRecipesLatest'
-import { UiImage } from '@/components/ui'
 import { type CardRecipeProps } from '@interfaces/CardRecipeProps.interface'
+import { UiImage } from '@shared-components/ui'
 
 export default function CardRecipe({
 	img,
