+++ tsconfig.json
@@ -32,6 +32,9 @@
       ],
       "@libs/*": [
         "./libs/*"
+      ],
+      "@shared-components/*": [
+        "./components/*"
       ]
     }
   },
